backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.77.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
limits==5.6.0
openai==2.16.0
orjson==3.11.7
packaging==24.2
pinecone==8.0.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.7
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.50.0
tiktoken==0.12.0
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.1
